Revert "Revert "Add Site Specific Settings (#3140)" (#3252)"
This reverts commit 3b2135e6cfa0c0064547e014315374a1e2b9350c.
