+++ Shared/UI/Util/ButtonStyles.swift
@@ -57,13 +57,14 @@ public struct NeevaButtonStyle: ButtonStyle {
     }
 
     let visualSpec: VisualSpec
+    var height: CGFloat = 48
     @Environment(\.isEnabled) private var isEnabled
 
     public func makeBody(configuration: Configuration) -> some View {
         configuration.label
-            .foregroundColor(visualSpec == VisualSpec.primary ? .white : .label)
+            .foregroundColor(visualSpec == VisualSpec.primary ? .white : .blue)
             .padding(.vertical, 8)
-            .frame(height: 48)
+            .frame(height: height)
             .background(
                 Capsule()
                     .fill(
@@ -79,7 +80,9 @@ public struct NeevaButtonStyle: ButtonStyle {
 }
 
 extension ButtonStyle where Self == NeevaButtonStyle {
-    public static func neeva(_ visualSpec: NeevaButtonStyle.VisualSpec) -> Self {
-        .init(visualSpec: visualSpec)
+    public static func neeva(_ visualSpec: NeevaButtonStyle.VisualSpec, height: CGFloat = 48)
+        -> Self
+    {
+        .init(visualSpec: visualSpec, height: height)
     }
 }
